Move alloc tests to test_core.py
